# backend/src/api-zod/generated/types/appSettings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Important Updates private task and chat API
 * OpenAPI spec version: 0.1.0
 */
import type { AppSettingsAutoLock } from './appSettingsAutoLock';
import type { AppSettingsBubbleStyle } from './appSettingsBubbleStyle';
import type { AppSettingsTheme } from './appSettingsTheme';

export interface AppSettings {
  theme: AppSettingsTheme;
  bubbleStyle: AppSettingsBubbleStyle;
  autoLock: AppSettingsAutoLock;
  notifications: boolean;
  journeyUrl: string;
}