# backend/src/api-zod/generated/types/device.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Important Updates private task and chat API
 * OpenAPI spec version: 0.1.0
 */

export interface Device {
  id: string;
  label: string;
  createdAt: Date;
  lastActiveAt: Date;
  current: boolean;
}