# backend/src/api-zod/generated/types/journey.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Important Updates private task and chat API
 * OpenAPI spec version: 0.1.0
 */

export interface Journey {
  title: string;
  description: string;
  url: string;
}