# backend/src/api-zod/generated/types/message.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Important Updates private task and chat API
 * OpenAPI spec version: 0.1.0
 */
import type { MessageDeliveryStatus } from './messageDeliveryStatus';

export interface Message {
  id: string;
  senderId: string;
  senderName: string;
  content: string;
  createdAt: Date;
  /** @nullable */
  editedAt: Date | null;
  /** @nullable */
  deletedAt: Date | null;
  deliveryStatus: MessageDeliveryStatus;
  /** @nullable */
  readAt: Date | null;
}