# backend/src/api-zod/generated/types/messageUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Important Updates private task and chat API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageUpdate {
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}