# backend/src/api-zod/generated/types/settingsUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Important Updates private task and chat API
 * OpenAPI spec version: 0.1.0
 */
import type { SettingsUpdateAutoLock } from './settingsUpdateAutoLock';
import type { SettingsUpdateBubbleStyle } from './settingsUpdateBubbleStyle';
import type { SettingsUpdateTheme } from './settingsUpdateTheme';

export interface SettingsUpdate {
  theme?: SettingsUpdateTheme;
  bubbleStyle?: SettingsUpdateBubbleStyle;
  autoLock?: SettingsUpdateAutoLock;
  notifications?: boolean;
  journeyUrl?: string;
}