# backend/src/api-zod/generated/types/task.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Important Updates private task and chat API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskStatus } from './taskStatus';

export interface Task {
  id: string;
  title: string;
  /** @nullable */
  description: string | null;
  status: TaskStatus;
  /** @nullable */
  assignedTo: string | null;
  createdBy: string;
  /** @nullable */
  dueDate: Date | null;
  createdAt: Date;
  updatedAt: Date;
}