# backend/src/api-zod/generated/types/taskSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Important Updates private task and chat API
 * OpenAPI spec version: 0.1.0
 */

export interface TaskSummary {
  total: number;
  pending: number;
  inProgress: number;
  complete: number;
}