# backend/src/api-zod/generated/types/taskUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Important Updates private task and chat API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskUpdateStatus } from './taskUpdateStatus';

export interface TaskUpdate {
  /**
     * @minLength 1
     * @maxLength 160
     */
  title?: string;
  /**
     * @maxLength 500
     * @nullable
     */
  description?: string | null;
  status?: TaskUpdateStatus;
  /** @nullable */
  assignedTo?: string | null;
  /** @nullable */
  dueDate?: Date | null;
}